deltete unused components and move things to keep a better project structure
