+++ mensasverige/features/account/screens/Settings.tsx
@@ -18,8 +18,8 @@ import { resetUserCredentials } from '../../common/services/authService';
 import ProfileEditAvatar from '../../common/components/ProfileEditAvatar';
 import { extractNumericValue } from '../../common/functions/extractNumericValue';
 import { MaterialIcons } from '@expo/vector-icons';
-import EditableField from '../../common/components/EditableField';
-import Dropdown, { DropdownOption } from '../../common/components/Dropdown';
+import EditableField from '../../common/components/inputs/EditableField';
+import Dropdown, { DropdownOption } from '../../common/components/inputs/Dropdown';
 import { Colors } from '@/constants/Colors';
 import { DEFAULT_SETTINGS } from '@/constants/DefaultSettings';
 import Slider from '@react-native-community/slider';
