--- mensasverige/features/common/components/NoWifiSpinner.tsx
@@ -1,7 +1,0 @@
-import React from 'react-native';
-import {ICustomTheme, Spinner, useTheme} from 'native-base';
-
-export const NoWifiSpinner = () => {
-  const theme = useTheme() as ICustomTheme;
-  return <Spinner size="sm" color={theme.colors.text[500]} />;
-};
