--- mensasverige/features/common/components/ShowSettingsLabelIcon.tsx
@@ -1,5 +1,0 @@
-const ShowSettingsLabelIconColor = (value: boolean | undefined): string => {
-  return value ? 'visibility' : 'visibility-off';
-};
-
-export default ShowSettingsLabelIconColor;
