+++ mensasverige/features/events/components/ParentEventDetails.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect } from 'react';
-import { LocationLinkButton } from '../../map/components/LocationLinkIcon';
+import { LocationLinkButton } from '@/features/map/components/LocationLinkIcon';
 import { ExternalRoot } from '../../../api_schema/types';
 import { fetchExternalRoot } from '../services/eventService';
 import { View } from 'react-native';
